Update integer types and add backpressure to block writer

+++ cache/cache.go
@@ -4,6 +4,6 @@ type Cache interface {
 	Add(key string, value []byte)
 	Get(key string) ([]byte, bool)
 	Remove(key string)
-	Len() int
+	Len() int64
 	Clear()
 }
