+++ fileio/mmap.go
@@ -19,7 +19,7 @@ type MMapIO struct {
 	mm   *MMap
 }
 
-func NewMMap(filename string, size int) *MMap {
+func NewMMap(filename string, size int64) *MMap {
 	mm := new(MMap)
 	mm.filename = filename
 	mm.open()
@@ -28,7 +28,7 @@ func NewMMap(filename string, size int) *MMap {
 	return mm
 }
 
-func (mm *MMap) resize(size int) {
+func (mm *MMap) resize(size int64) {
 	log.Error("Resizing: ", size)
 	err := syscall.Ftruncate(mm.fd, int64(size))
 	if err != nil {
@@ -46,23 +46,23 @@ func (mm *MMap) open() {
 	mm.file = f
 }
 
-func (mm *MMap) mmap(size int) {
+func (mm *MMap) mmap(size int64) {
 	log.Info("mmapping: ", size, int64(size))
-	data, err := syscall.Mmap(mm.fd, 0, size, syscall.PROT_WRITE|syscall.PROT_READ, syscall.MAP_PRIVATE)
+	data, err := syscall.Mmap(mm.fd, 0, int(size), syscall.PROT_WRITE|syscall.PROT_READ, syscall.MAP_PRIVATE)
 	if err != nil {
 		log.Error("Error mmapping: ", err)
 	}
 	mm.data = data
 }
 
-func (mm *MMap) extend(size int) {
+func (mm *MMap) extend(size int64) {
 	mm.file.Close()
 	mm.open()
 	mm.resize(size)
 	mm.mmap(size)
 }
 
-func NewMMapIO(filename string, size int) *MMapIO {
+func NewMMapIO(filename string, size int64) *MMapIO {
 	mmio := new(MMapIO)
 	mmio.mm = NewMMap(filename, size)
 	mmio.data = mmio.mm.GetData()
@@ -73,10 +73,10 @@ func (mm *MMap) GetData() []byte {
 	return mm.data
 }
 
-func (mmio *MMapIO) ReadAt(offset int, size int) []byte {
+func (mmio *MMapIO) ReadAt(offset int64, size int64) []byte {
 	return mmio.data[offset : offset+size]
 }
 
-func (mmio *MMapIO) WriteAt(offset int, b []byte) {
-	copy(mmio.data[offset:offset+len(b)], b)
+func (mmio *MMapIO) WriteAt(offset int64, b []byte) {
+	copy(mmio.data[offset:int(offset)+len(b)], b)
 }
